Update emscripten docker image to be based on emscripten 2.0.12 and boost 1.75.0.

+++ libsolutil/SwarmHash.cpp
@@ -29,7 +29,7 @@ using namespace solidity::util;
 namespace
 {
 
-bytes toLittleEndian(size_t _size)
+bytes toLittleEndian(uint64_t _size)
 {
 	bytes encoded(8);
 	for (size_t i = 0; i < 8; ++i)
